requirements.txt
google-adk>=0.1.0
google-generativeai>=0.3.0
tavily-python>=0.3.0
exa-py>=1.0.0
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
gunicorn